+++ src/views/Customers.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import CustomerForm from '../components/CustomerForm';
 import DataTable from '../components/DataTable';
-import Header from '../components/Header';
+import Header from '../elements/Header';
 import Modal from '../components/Modal';
 import useCustomers from '../hooks/useCustomers';
 import useCreateCustomer from '../hooks/useCreateCustomer';
